// src/app/models/DetailInteractionModel.ts
export type DetailInteractionModel = {
    interactionId: number
    interactionTypeId: number
    interactionType: string
    resource: string
    contactId: number
    contactName: string
    phone: string
    eMail: string
    mobil: any
    initialDate: string
    finalDate: string
    observations: string
    statusCode: string
    dinamycData?: []
  }